add: I'm gonna die

+++ backend/src/main/java/com/example/demo/service/ActivityService.java
@@ -3,9 +3,15 @@ package com.example.demo.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import javax.servlet.http.HttpServletRequest;
+
 import com.example.demo.model.Activity;
+
 import com.example.demo.repository.ActivityRepository;
-import com.example.demo.model.ResourceNotFoundException;
+
+import com.example.demo.service.JwtService;
+
+import com.example.demo.dto.ActivityResult;
 
 import java.util.List;
 
@@ -14,34 +20,85 @@ public class ActivityService {
     @Autowired
     private ActivityRepository activityRepository;
 
+    @Autowired
+    private JwtService jwtService;
+
     public List<Activity> getAllActivities() {
         return activityRepository.findAll();
     }
 
-    public Activity createActivity(Activity activity) {
-        return activityRepository.save(activity);
+    public ActivityResult createActivity(Activity activity, HttpServletRequest request) {
+
+        JwtResult jwtResult = jwtService.parseRequest(request);
+
+        ActivityResult result = new ActivityResult();
+        
+        if (!(jwtResult != null) && (!jwtResult.getPassed())) {
+            result.setResultCode(1);
+            return result;
+        }
+
+        result.setActivity(activityRepository.save(activity));
+        result.setResultCode(0);
+
+        return result;
     }
 
-    public void deleteActivity(Integer id) {
-        Activity activity = getActivity(id);
-        activityRepository.delete(activity);
+    public int deleteActivity(Integer id, HttpServletRequest request) {
+
+        JwtResult jwtResult = jwtService.parseRequest(request);
+        
+        if (!(jwtResult != null) && (!jwtResult.getPassed())) {
+            return 1;
+        }
+        
+        Activity activity = activityRepository.findById(id).orElse(null);
+        
+        User user = jwtResult.getUser();
+
+        if (user.getPermission() == 0) {
+            return 1;
+        } else if (activity == null) {
+            return 2;
+        } else {
+            activityRepository.delete(activity);
+            return 0;
+        }
     }
 
-    public Activity updateActivity(Integer id, Activity activityDetails) {
-        Activity activity = getActivity(id);
-        activity.setTitle(activityDetails.getTitle());
-        activity.setContent(activityDetails.getContent());
-        activity.setLocation(activityDetails.getLocation());
-        activity.setDateTime(activityDetails.getDateTime());
-        activity.setTags(activityDetails.getTags());
-        return activityRepository.save(activity);
+    public Activity updateActivity(Integer id, Activity activityDetails, HttpServletRequest request) {
+        
+        JwtResult jwtResult = jwtService.parseRequest(request);
+        ActivityResult result = new ActivityResult();
+        
+        if (!(jwtResult != null) && (!jwtResult.getPassed())) {
+            result.setResultCode(1);
+            return result;
+        }
+        
+        Activity activity = activityRepository.findById(id).orElse(null);
+
+        if (activity != null) {
+            if (activity.getId() == activityDetails.getId()) {
+                activity.setTitle(activityDetails.getTitle());
+                activity.setContent(activityDetails.getContent());
+                activity.setLocation(activityDetails.getLocation());
+                activity.setDateTime(activityDetails.getDateTime());
+                activity.setVisible(activityDetails.getVisible());
+                activityRepository.save(activity);
+                result.setResultCode(0);
+                result.setActivity(activity);
+            } else {
+                result.setResultCode(1);
+            }
+        } else {
+            result.setResultCode(2);
+        }
+
+        return result;
     }
 
     public Activity getActivityById(Integer id) {
-        return activityRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Activity", "id", id));
-    }
-
-    private Activity getActivity(Integer id) {
-        return activityRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Activity", "id", id));
+        return activityRepository.findById(id).orElse(null);
     }
 }
